requirements.txt
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
pydantic>=2.6.0
pydantic-settings>=2.2.0
opencv-python-headless>=4.9.0
pillow>=10.2.0
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
ultralytics>=8.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
requests>=2.32.0
PyYAML>=6.0.1
loguru>=0.7.2